+++ support/src/traits.rs
@@ -69,7 +69,9 @@ pub trait GenerateBenchmarkOrigin<OuterOrigin, AccountId, SubjectId> {
 	fn generate_origin(sender: AccountId, subject: SubjectId) -> OuterOrigin;
 }
 
+/// Trait that allows types to implement a default value for a given length,
+/// only when running benchmarks.
 #[cfg(feature = "runtime-benchmarks")]
 pub trait DefaultForLength {
-	fn get_default(length: u32) -> Self;
+	fn get_default(length: usize) -> Self;
 }
